refactor: replace HTML parsing with Daraz AJAX API

+++ src/daraz_scraper/models.py
@@ -7,5 +7,5 @@ class Product:
     name: str
     price: float
     sold: int
-    rating: int
+    rating: float
     link: str
